permission handling

+++ src/main/java/com/example/demowebshop/controller/ProductController.java
@@ -2,7 +2,9 @@ package com.example.demowebshop.controller;
 
 import com.example.demowebshop.Util.RetryCommand;
 import com.example.demowebshop.model.Product;
+import com.example.demowebshop.model.User;
 import com.example.demowebshop.service.IProductService;
+import com.example.demowebshop.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
@@ -14,6 +16,9 @@ public class ProductController {
 
     @Autowired
     private IProductService productService;
+
+    @Autowired
+    private UserService userService;
 
     private RetryCommand<Product> retryCommand;
 
@@ -28,8 +33,13 @@ public class ProductController {
     }
 
     @RequestMapping("/products/delete/{id}")
-    private void deleteProduct(@PathVariable("id") int id) {
-        productService.delete(id);
+    private void deleteProduct(@PathVariable("id") int id, int userId) {
+        User user = userService.getUserById(userId);
+        if(user.isPermitted("deleteProduct")) {
+            productService.delete(id);
+        } else {
+            System.out.println("Don't have permission");
+        }
     }
 
     @RequestMapping("/product")
